requirements.txt
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.21
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
